v1.3
nav bar is now responsive using css.  would like the bar to stay at the
top of the page on next iteration.  perhaps use of js to hide the nav
bar when scrolling down and show again when scrolling up.  need to
still the website and iPhone icons per assignment requirement.  not
completely satisfied with the placement of the login and sign up links
@ all viewpoints.